Move `RequestSession`, `DownloadSession` and `UploadSession` to separate files
