+++ Source/Session.swift
@@ -24,243 +24,6 @@
 
 import Foundation
 
-public protocol RequestSession {
-    func request(_ url: URLConvertible,
-                 method: HTTPMethod,
-                 parameters: Parameters?,
-                 encoding: ParameterEncoding,
-                 headers: HTTPHeaders?,
-                 interceptor: RequestInterceptor?) -> DataRequest
-
-    func request<Parameters: Encodable>(_ url: URLConvertible,
-                                        method: HTTPMethod,
-                                        parameters: Parameters?,
-                                        encoder: ParameterEncoder,
-                                        headers: HTTPHeaders?,
-                                        interceptor: RequestInterceptor?) -> DataRequest
-
-    func request(_ convertible: URLRequestConvertible, interceptor: RequestInterceptor?) -> DataRequest
-}
-
-extension RequestSession {
-    public func request(_ url: URLConvertible,
-                        method: HTTPMethod = .get,
-                        parameters: Parameters? = nil,
-                        encoding: ParameterEncoding = URLEncoding.default,
-                        headers: HTTPHeaders? = nil,
-                        interceptor: RequestInterceptor? = nil) -> DataRequest {
-        let convertible = RequestConvertible(url: url,
-                                             method: method,
-                                             parameters: parameters,
-                                             encoding: encoding,
-                                             headers: headers)
-
-        return request(convertible, interceptor: interceptor)
-    }
-
-    public func request<Parameters: Encodable>(_ url: URLConvertible,
-                                               method: HTTPMethod = .get,
-                                               parameters: Parameters? = nil,
-                                               encoder: ParameterEncoder = JSONParameterEncoder.default,
-                                               headers: HTTPHeaders? = nil,
-                                               interceptor: RequestInterceptor? = nil) -> DataRequest {
-        let convertible = RequestEncodableConvertible(url: url,
-                                                      method: method,
-                                                      parameters: parameters,
-                                                      encoder: encoder,
-                                                      headers: headers)
-
-        return request(convertible, interceptor: interceptor)
-    }
-}
-
-public protocol DownloadSession {
-    func download(_ convertible: URLConvertible,
-                  method: HTTPMethod,
-                  parameters: Parameters?,
-                  encoding: ParameterEncoding,
-                  headers: HTTPHeaders?,
-                  interceptor: RequestInterceptor?,
-                  to destination: DownloadRequest.Destination?) -> DownloadRequest
-
-    func download<Parameters: Encodable>(_ convertible: URLConvertible,
-                                         method: HTTPMethod,
-                                         parameters: Parameters?,
-                                         encoder: ParameterEncoder,
-                                         headers: HTTPHeaders?,
-                                         interceptor: RequestInterceptor?,
-                                         to destination: DownloadRequest.Destination?) -> DownloadRequest
-
-    func download(_ convertible: URLRequestConvertible,
-                  interceptor: RequestInterceptor?,
-                  to destination: DownloadRequest.Destination?) -> DownloadRequest
-
-    func download(resumingWith data: Data,
-                  interceptor: RequestInterceptor?,
-                  to destination: DownloadRequest.Destination?) -> DownloadRequest
-}
-
-extension DownloadSession {
-    public func download(_ convertible: URLConvertible,
-                         method: HTTPMethod = .get,
-                         parameters: Parameters? = nil,
-                         encoding: ParameterEncoding = URLEncoding.default,
-                         headers: HTTPHeaders? = nil,
-                         interceptor: RequestInterceptor? = nil,
-                         to destination: DownloadRequest.Destination? = nil) -> DownloadRequest {
-        let convertible = RequestConvertible(url: convertible,
-                                             method: method,
-                                             parameters: parameters,
-                                             encoding: encoding,
-                                             headers: headers)
-
-        return download(convertible, interceptor: interceptor, to: destination)
-    }
-
-    public func download<Parameters: Encodable>(_ convertible: URLConvertible,
-                                                method: HTTPMethod = .get,
-                                                parameters: Parameters? = nil,
-                                                encoder: ParameterEncoder = JSONParameterEncoder.default,
-                                                headers: HTTPHeaders? = nil,
-                                                interceptor: RequestInterceptor? = nil,
-                                                to destination: DownloadRequest.Destination? = nil) -> DownloadRequest {
-        let convertible = RequestEncodableConvertible(url: convertible,
-                                                      method: method,
-                                                      parameters: parameters,
-                                                      encoder: encoder,
-                                                      headers: headers)
-
-        return download(convertible, interceptor: interceptor, to: destination)
-    }
-}
-
-public protocol UploadSession {
-    func upload(_ data: Data,
-                to convertible: URLConvertible,
-                method: HTTPMethod,
-                headers: HTTPHeaders?,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(_ data: Data,
-                with convertible: URLRequestConvertible,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(_ fileURL: URL,
-                to convertible: URLConvertible,
-                method: HTTPMethod,
-                headers: HTTPHeaders?,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(_ fileURL: URL,
-                with convertible: URLRequestConvertible,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(_ stream: InputStream,
-                to convertible: URLConvertible,
-                method: HTTPMethod,
-                headers: HTTPHeaders?,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(_ stream: InputStream,
-                with convertible: URLRequestConvertible,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(multipartFormData: @escaping (MultipartFormData) -> Void,
-                usingThreshold encodingMemoryThreshold: UInt64,
-                fileManager: FileManager,
-                to url: URLConvertible,
-                method: HTTPMethod,
-                headers: HTTPHeaders?,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(multipartFormData: MultipartFormData,
-                usingThreshold encodingMemoryThreshold: UInt64,
-                with request: URLRequestConvertible,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(_ uploadable: UploadRequest.Uploadable,
-                with convertible: URLRequestConvertible,
-                interceptor: RequestInterceptor?) -> UploadRequest
-
-    func upload(_ upload: UploadConvertible, interceptor: RequestInterceptor?) -> UploadRequest
-}
-
-extension UploadSession {
-    public func upload(_ data: Data,
-                       to convertible: URLConvertible,
-                       method: HTTPMethod = .post,
-                       headers: HTTPHeaders? = nil,
-                       interceptor: RequestInterceptor? = nil) -> UploadRequest {
-        let convertible = ParameterlessRequestConvertible(url: convertible, method: method, headers: headers)
-
-        return upload(data, with: convertible, interceptor: interceptor)
-    }
-
-    public func upload(_ data: Data,
-                       with convertible: URLRequestConvertible,
-                       interceptor: RequestInterceptor? = nil) -> UploadRequest {
-        return upload(.data(data), with: convertible, interceptor: interceptor)
-    }
-
-    public func upload(_ fileURL: URL,
-                       to convertible: URLConvertible,
-                       method: HTTPMethod = .post,
-                       headers: HTTPHeaders? = nil,
-                       interceptor: RequestInterceptor? = nil) -> UploadRequest {
-        let convertible = ParameterlessRequestConvertible(url: convertible, method: method, headers: headers)
-
-        return upload(fileURL, with: convertible, interceptor: interceptor)
-    }
-
-    public func upload(_ fileURL: URL,
-                       with convertible: URLRequestConvertible,
-                       interceptor: RequestInterceptor? = nil) -> UploadRequest {
-        return upload(.file(fileURL, shouldRemove: false), with: convertible, interceptor: interceptor)
-    }
-
-    public func upload(_ stream: InputStream,
-                       to convertible: URLConvertible,
-                       method: HTTPMethod = .post,
-                       headers: HTTPHeaders? = nil,
-                       interceptor: RequestInterceptor? = nil) -> UploadRequest {
-        let convertible = ParameterlessRequestConvertible(url: convertible, method: method, headers: headers)
-
-        return upload(stream, with: convertible, interceptor: interceptor)
-    }
-
-    public func upload(_ stream: InputStream,
-                       with convertible: URLRequestConvertible,
-                       interceptor: RequestInterceptor? = nil) -> UploadRequest {
-        return upload(.stream(stream), with: convertible, interceptor: interceptor)
-    }
-
-    public func upload(multipartFormData: @escaping (MultipartFormData) -> Void,
-                       usingThreshold encodingMemoryThreshold: UInt64 = MultipartFormData.encodingMemoryThreshold,
-                       fileManager: FileManager = .default,
-                       to url: URLConvertible,
-                       method: HTTPMethod = .post,
-                       headers: HTTPHeaders? = nil,
-                       interceptor: RequestInterceptor? = nil) -> UploadRequest {
-        let convertible = ParameterlessRequestConvertible(url: url, method: method, headers: headers)
-
-        let formData = MultipartFormData(fileManager: fileManager)
-        multipartFormData(formData)
-
-        return upload(multipartFormData: formData,
-                      usingThreshold: encodingMemoryThreshold,
-                      with: convertible,
-                      interceptor: interceptor)
-    }
-
-    public func upload(_ uploadable: UploadRequest.Uploadable,
-                       with convertible: URLRequestConvertible,
-                       interceptor: RequestInterceptor?) -> UploadRequest {
-        let uploadable = Upload(request: convertible, uploadable: uploadable)
-
-        return upload(uploadable, interceptor: interceptor)
-    }
-}
-
 public protocol SessionProtocol: RequestSession, DownloadSession, UploadSession {
     var session: URLSession { get }
     var delegate: SessionDelegate { get }
@@ -309,30 +72,6 @@ public struct RequestConvertible: URLRequestConvertible {
         return try encoding.encode(request, with: parameters)
     }
 }
-
-public struct ParameterlessRequestConvertible: URLRequestConvertible {
-    let url: URLConvertible
-    let method: HTTPMethod
-    let headers: HTTPHeaders?
-
-    public func asURLRequest() throws -> URLRequest {
-        return try URLRequest(url: url, method: method, headers: headers)
-    }
-}
-
-public struct Upload: UploadConvertible {
-    let request: URLRequestConvertible
-    let uploadable: UploadableConvertible
-
-    public func createUploadable() throws -> UploadRequest.Uploadable {
-        return try uploadable.createUploadable()
-    }
-
-    public func asURLRequest() throws -> URLRequest {
-        return try request.asURLRequest()
-    }
-}
-
 
 open class Session: SessionProtocol {
     public static let `default` = Session()
